Split ABI Encoder/Decoder tests into separate files
