+++ packages/utils/src/abi_encoder/evm_data_types/address.ts
@@ -13,7 +13,7 @@ export class Address extends PayloadDataType {
     private static readonly _SIZE_KNOWN_AT_COMPILE_TIME: boolean = true;
     private static readonly _ADDRESS_SIZE_IN_BYTES = 20;
     private static readonly _DECODED_ADDRESS_OFFSET_IN_BYTES = Constants.EVM_WORD_WIDTH_IN_BYTES -
-        Address._ADDRESS_SIZE_IN_BYTES;
+    Address._ADDRESS_SIZE_IN_BYTES;
 
     public static matchType(type: string): boolean {
         return type === 'address';
