+++ packages/utils/src/abi_encoder/evm_data_types/static_bytes.ts
@@ -46,7 +46,7 @@ export class StaticBytes extends PayloadDataType {
         if (valueBuf.byteLength > this._width) {
             throw new Error(
                 `Tried to assign ${value} (${
-                valueBuf.byteLength
+                    valueBuf.byteLength
                 } bytes), which exceeds max bytes that can be stored in a ${this.getSignature()}`,
             );
         } else if (value.length % 2 !== 0) {
